updated notes

+++ hardhat/ethers-simple-storage/SimpleStorage.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.18;
 
 contract SimpleStorage {
-    uint16 myNumber;
+    uint256 myNumber;
 
-    function storeNumber(uint16 number) public {
+    function storeNumber(uint256 number) public {
         myNumber = number;
     }
 
-    function getNumber() public view returns (uint16) {
+    function getNumber() public view returns (uint256) {
         return myNumber;
     }
 }
